Add Tesseract API support + Clean up the project
