--- Login/GettingStarted.Xamarin
@@ -1,4 +1,0 @@
-<GettingStarted>
-	<LocalContent>GS\Android\CS\AndroidApp\GettingStarted.html</LocalContent>
-	<EmbeddedNavigation>false</EmbeddedNavigation>
-</GettingStarted>
