+++ Login/packages.config
@@ -10,4 +10,5 @@
   <package id="Xamarin.Android.Support.v7.AppCompat" version="25.1.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.Vector.Drawable" version="25.1.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Build.Download" version="0.4.2" targetFramework="monoandroid71" />
+  <package id="Xamarin.Tesseract" version="0.3.2" targetFramework="monoandroid50" />
 </packages>
